# swagger/v1/swagger.yaml
...
openapi: 3.0.1
info:
  title: API V1
  version: v1
paths:
...
  "/v1/vouchers":
    get:
      summary: Lists vouchers
      tags:
      - Vouchers
      security:
      - BearerAuth: []
      parameters:
      - name: Authorization
        in: header
        required: true
        description: Bearer JWT token
        schema:
          type: string
      - name: vendor_id
        in: query
        required: false
        description: Filter by vendor ID
        schema:
          type: integer
      - name: event_id
        in: query
        required: false
        description: Filter by event ID
        schema:
          type: integer
      responses:
        '200':
          description: Vouchers filtered by event_id
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        title:
                          type: string
                          example: 20% Off Deal
                        description:
                          type: string
                          nullable: true
                          example: Get 20% off on all items
                        voucher_uuid:
                          type: string
                          format: uuid
                        voucher_code:
                          type: string
                          example: SAVE20
                        status:
                          type: string
                          example: active
                        vendor_id:
                          type: integer
                        event_id:
                          type: integer
                        start_date:
                          type: string
                          format: date
                          example: '2024-01-01'
                        end_date:
                          type: string
                          format: date
                          example: '2024-12-31'
                        start_time:
                          type: string
                          format: time
                          example: '00:00:00'
                        end_time:
                          type: string
                          format: time
                          example: '23:59:59'
                        total_redemption_available:
                          type: integer
                          nullable: true
                          example: 100
                        is_unlimited:
                          type: boolean
                          example: false
                          description: If true, voucher has no redemption quota limit
                        redeemed_count:
                          type: integer
                          example: 0
                        max_redemptions_per_user:
                          type: integer
                          nullable: true
                          example: 5
                        voucher_type:
                          type: string
                          enum:
                          - fixed_amount
                          - percentage
                          - free_item
                          example: percentage
                        voucher_value:
                          type: string
                          example: '20.00'
                        image_path:
                          type: string
                          nullable: true
                          example: https://example.com/images/voucher.jpg
                        created_at:
                          type: string
                          format: date_time
                        updated_at:
                          type: string
                          format: date_time
                      required:
                      - id
                      - title
                      - voucher_uuid
                      - vendor_id
                      - event_id
                    example:
                    - id: 1
                      title: V1E1
                      voucher_uuid: 550e8400-e29b-41d4-a716-446655440000
                      vendor_id: 1
                      event_id: 1
                      voucher_code: SAVE20
                      status: active
                      voucher_type: percentage
                      voucher_value: '20.00'
                required:
                - success
                - message
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized
    post:
      summary: Creates a new voucher
      tags:
      - Vouchers
      security:
      - BearerAuth: []
      parameters:
      - name: Authorization
        in: header
        required: true
        description: Bearer JWT token
        schema:
          type: string
      responses:
        '201':
          description: Voucher created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Success
                  data:
                    type: object
                required:
                - success
                - message
        '403':
          description: Forbidden - Cannot create voucher for another vendor
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                  errors:
                    type: array
                    items:
                      type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Voucher title
                description:
                  type: string
                  description: Voucher description
                vendor_id:
                  type: integer
                  description: Vendor ID
                event_id:
                  type: integer
                  description: Event ID
                voucher_code:
                  type: string
                  description: Voucher code
                status:
                  type: string
                  enum: [active, inactive]
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                start_time:
                  type: string
                  format: time
                end_time:
                  type: string
                  format: time
                total_redemption_available:
                  type: integer
                  nullable: true
                  description: Total quota (ignored if is_unlimited is true)
                is_unlimited:
                  type: boolean
                  default: false
                  description: If true, voucher has no redemption quota limit
                max_redemptions_per_user:
                  type: integer
                  nullable: true
                voucher_type:
                  type: string
                  enum: [fixed_amount, percentage, free_item]
                voucher_value:
                  type: number
                voucher_category:
                  type: string
                image:
                  type: string
                  format: binary
                  description: Voucher image file
              required:
                - title
                - vendor_id
                - event_id
        required: true
        description: Voucher data
  "/v1/vouchers/{id}":
    parameters:
    - name: id
      in: path
      required: true
      description: Voucher ID
      schema:
        type: integer
    get:
      summary: Retrieves a voucher
      tags:
      - Vouchers
      security:
      - BearerAuth: []
      parameters:
      - name: Authorization
        in: header
        required: true
        description: Bearer JWT token
        schema:
          type: string
      responses:
        '200':
          description: Voucher retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Success
                  data:
                    type: object
                required:
                - success
                - message
        '403':
          description: Forbidden - Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: You are not authorized to show? this voucher.
        '404':
          description: Voucher not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Resource not found
    patch:
      summary: Updates a voucher
      tags:
      - Vouchers
      security:
      - BearerAuth: []
      parameters:
      - name: Authorization
        in: header
        required: true
        description: Bearer JWT token
        schema:
          type: string
      responses:
        '200':
          description: Voucher updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Success
                  data:
                    type: object
                required:
                - success
                - message
        '403':
          description: Forbidden - Cannot update another vendor's voucher
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: string
    delete:
      summary: Deletes a voucher
      tags:
      - Vouchers
      security:
      - BearerAuth: []
      parameters:
      - name: Authorization
        in: header
        required: true
        description: Bearer JWT token
        schema:
          type: string
      responses:
        '204':
          description: Voucher deleted successfully
        '403':
          description: Forbidden - Cannot delete another vendor's voucher
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
        '404':
          description: Voucher not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Resource not found
servers:
- url: https://{defaultHost}
  variables:
    defaultHost:
      default: www.example.com
- url: http://localhost:3001
  description: Local Development